+++ .coveragerc
@@ -1,2 +1,2 @@
 [run]
-omit = .venv/*,config/wsgi.py,manage.py,**/tests/*
+omit = .venv/*,conf/wsgi.py,manage.py,**/tests/*
