+++ docker/cmd-webserver.sh
@@ -1,3 +1,3 @@
 #!/bin/bash -xe
 
-gunicorn config.wsgi --bind 0.0.0.0:$PORT
+gunicorn conf.wsgi --bind 0.0.0.0:$PORT
