+++ proxy/core/views.py
@@ -1,0 +1,9 @@
+from proxy.utils import BaseProxyView
+
+
+class RobotsView(BaseProxyView):
+    pass
+
+
+class SitemapView(BaseProxyView):
+    pass
